+++ packages/examples/packages/send-flow/src/components/SendForm.tsx
@@ -6,13 +6,13 @@ import {
   Icon,
   Image,
   Input,
+  AddressInput,
   Text,
   type SnapComponent,
 } from '@metamask/snaps-sdk/jsx';
 
 import { AccountSelector } from './AccountSelector';
 import btcIcon from '../images/btc.svg';
-import jazzicon3 from '../images/jazzicon3.svg';
 import type { Account, SendFormErrors } from '../types';
 
 /**
@@ -22,16 +22,14 @@ import type { Account, SendFormErrors } from '../types';
  * @property accounts - The available accounts.
  * @property errors - The form errors.
  * @property selectedCurrency - The selected currency to display.
- * @property displayClearIcon - Whether to display the clear icon or not.
- * @property flushToAddress - Whether to flush the address field or not.
+ * @property displayAvatar - Whether to display the avatar of the address.
  */
 export type SendFormProps = {
   selectedAccount: string;
   accounts: Account[];
   errors?: SendFormErrors;
   selectedCurrency: 'BTC' | '$';
-  displayClearIcon: boolean;
-  flushToAddress?: boolean;
+  displayAvatar?: boolean | undefined;
 };
 
 /**
@@ -42,8 +40,7 @@ export type SendFormProps = {
  * @param props.accounts - The available accounts.
  * @param props.errors - The form errors.
  * @param props.selectedCurrency - The selected currency to display.
- * @param props.displayClearIcon - Whether to display the clear icon or not.
- * @param props.flushToAddress - Whether to flush the address field or not.
+ * @param props.displayAvatar - Whether to display the avatar of the address.
  * @returns The SendForm component.
  */
 export const SendForm: SnapComponent<SendFormProps> = ({
@@ -51,8 +48,7 @@ export const SendForm: SnapComponent<SendFormProps> = ({
   accounts,
   errors,
   selectedCurrency,
-  displayClearIcon,
-  flushToAddress,
+  displayAvatar,
 }) => (
   <Form name="sendForm">
     <AccountSelector selectedAccount={selectedAccount} accounts={accounts} />
@@ -69,21 +65,12 @@ export const SendForm: SnapComponent<SendFormProps> = ({
       </Box>
     </Field>
     <Field label="To account" error={errors?.to}>
-      <Box>
-        <Image src={jazzicon3} />
-      </Box>
-      <Input
+      <AddressInput
         name="to"
+        chainId="eip155:0"
         placeholder="Enter receiving address"
-        value={flushToAddress ? '' : undefined}
+        displayAvatar={displayAvatar}
       />
-      {displayClearIcon && (
-        <Box>
-          <Button name="clear">
-            <Icon name="close" color="primary" />
-          </Button>
-        </Box>
-      )}
     </Field>
   </Form>
 );