+++ packages/snaps-sdk/src/jsx/components/form/Field.ts
@@ -1,11 +1,13 @@
+import type { AddressInputElement } from './AddressInput';
+import type { AssetSelectorElement } from './AssetSelector';
 import type { CheckboxElement } from './Checkbox';
 import type { DropdownElement } from './Dropdown';
 import type { FileInputElement } from './FileInput';
 import type { InputElement } from './Input';
 import type { RadioGroupElement } from './RadioGroup';
 import type { SelectorElement } from './Selector';
-import { createSnapComponent } from '../../component';
 import type { GenericSnapChildren } from '../../component';
+import { createSnapComponent } from '../../component';
 
 /**
  * The props of the {@link Field} component.
@@ -26,7 +28,9 @@ export type FieldProps = {
     | FileInputElement
     | InputElement
     | CheckboxElement
-    | SelectorElement;
+    | SelectorElement
+    | AssetSelectorElement
+    | AddressInputElement;
 };
 
 const TYPE = 'Field';