+++ packages/examples/packages/webpack-plugin/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.4]
+
+### Fixed
+
+- Bump `test-snaps` version ([#3192](https://github.com/MetaMask/snaps/pull/3192))
+
 ## [2.1.3]
 
 ### Fixed
@@ -65,7 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - The version of the package no longer needs to match the version of all other
     MetaMask Snaps packages.
 
-[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/webpack-plugin-example-snap@2.1.3...HEAD
+[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/webpack-plugin-example-snap@2.1.4...HEAD
+[2.1.4]: https://github.com/MetaMask/snaps/compare/@metamask/webpack-plugin-example-snap@2.1.3...@metamask/webpack-plugin-example-snap@2.1.4
 [2.1.3]: https://github.com/MetaMask/snaps/compare/@metamask/webpack-plugin-example-snap@2.1.2...@metamask/webpack-plugin-example-snap@2.1.3
 [2.1.2]: https://github.com/MetaMask/snaps/compare/@metamask/webpack-plugin-example-snap@2.1.1...@metamask/webpack-plugin-example-snap@2.1.2
 [2.1.1]: https://github.com/MetaMask/snaps/compare/@metamask/webpack-plugin-example-snap@2.1.0...@metamask/webpack-plugin-example-snap@2.1.1